Fix ClientClass and main function which cause error in master Build
