+++ DoLah/DoLah.vcxproj
@@ -123,7 +123,7 @@
     <ClCompile Include="basic.cpp" />
     <ClCompile Include="CommandInvoker.cpp" />
     <ClCompile Include="Models\Calendar.cpp" />
-    <ClCompile Include="Models\CalendarHistory.cpp" />
+    <ClCompile Include="Models\CommandHistory.cpp" />
     <ClCompile Include="Commands\AddTaskCommand.cpp" />
     <ClCompile Include="Commands\ClearTaskCommand.cpp" />
     <ClCompile Include="Commands\Command.cpp" />
@@ -141,7 +141,7 @@
     <ClInclude Include="basic.h" />
     <ClInclude Include="CommandInvoker.h" />
     <ClInclude Include="Models\Calendar.h" />
-    <ClInclude Include="Models\CalendarHistory.h" />
+    <ClInclude Include="Models\CommandHistory.h" />
     <ClInclude Include="Commands\Command.h" />
     <ClInclude Include="DoLahClient.h" />
     <ClInclude Include="Models\Task.h" />
