+++ DoLah/DoLah.vcxproj.filters
@@ -3,7 +3,6 @@
   <ItemGroup>
     <ClCompile Include="basic.cpp" />
     <ClCompile Include="Models\Calendar.cpp" />
-    <ClCompile Include="Models\CalendarHistory.cpp" />
     <ClCompile Include="Commands\AddTaskCommand.cpp" />
     <ClCompile Include="Commands\ClearTaskCommand.cpp" />
     <ClCompile Include="Commands\Command.cpp" />
@@ -17,16 +16,17 @@
     <ClCompile Include="Models\EventTask.cpp" />
     <ClCompile Include="Models\Task.cpp" />
     <ClCompile Include="CommandInvoker.cpp" />
+    <ClCompile Include="Models\CommandHistory.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="basic.h" />
     <ClInclude Include="Models\Calendar.h" />
-    <ClInclude Include="Models\CalendarHistory.h" />
     <ClInclude Include="Commands\Command.h" />
     <ClInclude Include="DoLahClient.h" />
     <ClInclude Include="Models\Task.h" />
     <ClInclude Include="Models\TaskFactory.h" />
     <ClInclude Include="CommandInvoker.h" />
+    <ClInclude Include="Models\CommandHistory.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
