--- DoLah/Models/CalendarHistory.cpp
@@ -1,22 +1,0 @@
-#include "CalendarHistory.h"
-
-namespace DoLah {
-	//public
-	CalendarHistory::CalendarHistory() {
-
-	}
-
-	CalendarHistory::~CalendarHistory() {
-
-	}
-
-	void CalendarHistory::redo() {
-		//TODO
-	}
-
-	void CalendarHistory::undo() {
-		//TODO
-	}
-
-	//private
-}
