+++ DoLah/Models/CommandHistory.cpp
@@ -1,0 +1,22 @@
+#include "CommandHistory.h"
+
+namespace DoLah {
+	//public
+    CommandHistory::CommandHistory() {
+
+	}
+
+    CommandHistory::~CommandHistory() {
+
+	}
+
+	void CommandHistory::redo() {
+		//TODO
+	}
+
+	void CommandHistory::undo() {
+		//TODO
+	}
+
+	//private
+}
