+++ DoLahUI/DoLahCLI.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <string>
+#include <iostream>
+#include "DoLahClient.h"
+
+void startConsole();
