+++ DoLahUI/main.cpp
@@ -1,20 +1,17 @@
 #include "DoLahUI.h"
+#include "DoLahCLI.h"
 
-namespace DoLah {
 
-    int main(int argc, char *argv[])
-    {
-        if (argc < 2) {
-            QApplication app(argc, argv);
-            DoLahUI window;
-            window.show();
-            FreeConsole();
-            return app.exec();
-        }
-        else {
-            startConsole();
-            return 0;
-        }
+int main(int argc, char *argv[])
+{
+    if (argc < 2) {
+        QApplication app(argc, argv);
+        DoLah::DoLahUI window;
+        window.show();
+        FreeConsole();
+        return app.exec();
+    } else {
+        startConsole();
+        return 0;
     }
-
 }
